# Swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Sistema de Subastas API",
      "description": "API REST del Sistema de Subastas — UADE Desarrollo de Aplicaciones I, 1C 2026. Gestiona registración de postores, catálogo, subastas, pujas, ventas, multas, seguros y solicitudes de incorporación de bienes.",
      "contact": {
        "name": "Equipo DA1",
        "email": "agustin.torres@majily.com"
      },
      "version": "0.1.0"
    },
    "paths": {
...
      "/usuarios/registro/etapa-1": {
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "Registrar datos iniciales del postor (etapa 1)",
          "operationId": "registrar_etapa_1_usuarios_registro_etapa_1_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRegistroEtapa1"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{usuario_id}/aprobacion": {
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "La empresa aprueba y asigna categoría tras la verificación externa",
          "operationId": "aprobar_usuario_usuarios__usuario_id__aprobacion_post",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioAprobacion"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{usuario_id}/registro/etapa-2": {
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "Postor completa el registro y genera su clave (etapa 2)",
          "operationId": "registrar_etapa_2_usuarios__usuario_id__registro_etapa_2_post",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioRegistroEtapa2"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/usuarios/{usuario_id}": {
        "get": {
          "tags": [
            "Usuarios"
          ],
          "summary": "Obtener usuario por id",
          "operationId": "obtener_usuario_usuarios__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/usuarios/{usuario_id}/cuentas-cobro": {
        "post": {
          "tags": [
            "Usuarios"
          ],
          "summary": "Declarar cuenta a la vista para liquidar ventas",
          "operationId": "crear_cuenta_cobro_usuarios__usuario_id__cuentas_cobro_post",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuentaCobroCreate"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/CuentaCobroOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "get": {
          "tags": [
            "Usuarios"
          ],
          "summary": "Listar cuentas de cobro del usuario",
          "operationId": "listar_cuentas_cobro_usuarios__usuario_id__cuentas_cobro_get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/CuentaCobroOut"
                    },
                    "title": "Response Listar Cuentas Cobro Usuarios  Usuario Id  Cuentas Cobro Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/pujas/usuario/{usuario_id}": {
        "get": {
          "tags": [
            "Pujas"
          ],
          "summary": "Historial de pujas de un usuario",
          "operationId": "historial_usuario_pujas_usuario__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PujaOut"
                    },
                    "title": "Response Historial Usuario Pujas Usuario  Usuario Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/multas/usuario/{usuario_id}": {
        "get": {
          "tags": [
            "Multas"
          ],
          "summary": "Listar Por Usuario",
          "operationId": "listar_por_usuario_multas_usuario__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MultaOut"
                    },
                    "title": "Response Listar Por Usuario Multas Usuario  Usuario Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/metricas/usuario/{usuario_id}": {
        "get": {
          "tags": [
            "Métricas"
          ],
          "summary": "Métricas de participación del usuario",
          "operationId": "metricas_usuario_metricas_usuario__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "CategoriaUsuario": {
          "type": "string",
          "enum": [
            "comun",
            "especial",
            "plata",
            "oro",
            "platino"
          ],
          "title": "CategoriaUsuario"
        },
        "CuentaCobroCreate": {
          "properties": {
            "banco": {
              "type": "string",
              "title": "Banco"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            },
            "numero_cuenta": {
              "type": "string",
              "title": "Numero Cuenta"
            },
            "titular": {
              "type": "string",
              "title": "Titular"
            },
            "declarada_antes_subasta": {
              "type": "boolean",
              "title": "Declarada Antes Subasta",
              "default": false
            }
          },
          "type": "object",
          "required": [
            "banco",
            "pais",
            "numero_cuenta",
            "titular"
          ],
          "title": "CuentaCobroCreate"
        },
        "CuentaCobroOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "banco": {
              "type": "string",
              "title": "Banco"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            },
            "numero_cuenta": {
              "type": "string",
              "title": "Numero Cuenta"
            },
            "titular": {
              "type": "string",
              "title": "Titular"
            },
            "declarada_antes_subasta": {
              "type": "boolean",
              "title": "Declarada Antes Subasta"
            }
          },
          "type": "object",
          "required": [
            "id",
            "usuario_id",
            "banco",
            "pais",
            "numero_cuenta",
            "titular",
            "declarada_antes_subasta"
          ],
          "title": "CuentaCobroOut"
        },
...
        "EstadoPuja": {
          "type": "string",
          "enum": [
            "pendiente",
            "confirmada",
            "rechazada"
          ],
          "title": "EstadoPuja"
        },
        "EstadoRegistro": {
          "type": "string",
          "enum": [
            "pendiente_verificacion",
            "aprobado_fase_1",
            "completo",
            "rechazado",
            "bloqueado"
          ],
          "title": "EstadoRegistro"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "MultaOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "venta_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Venta Id"
            },
            "monto": {
              "type": "string",
              "title": "Monto"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda"
            },
            "motivo": {
              "type": "string",
              "title": "Motivo"
            },
            "pagada": {
              "type": "boolean",
              "title": "Pagada"
            },
            "fecha": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha"
            }
          },
          "type": "object",
          "required": [
            "id",
            "usuario_id",
            "venta_id",
            "monto",
            "moneda",
            "motivo",
            "pagada",
            "fecha"
          ],
          "title": "MultaOut"
        },
...
        "PujaOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "subasta_id": {
              "type": "integer",
              "title": "Subasta Id"
            },
            "catalogo_item_id": {
              "type": "integer",
              "title": "Catalogo Item Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "monto": {
              "type": "string",
              "title": "Monto"
            },
            "fecha_hora": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha Hora"
            },
            "estado": {
              "$ref": "#/components/schemas/EstadoPuja"
            }
          },
          "type": "object",
          "required": [
            "id",
            "subasta_id",
            "catalogo_item_id",
            "usuario_id",
            "monto",
            "fecha_hora",
            "estado"
          ],
          "title": "PujaOut"
        },
...
        "UsuarioAprobacion": {
          "properties": {
            "categoria": {
              "$ref": "#/components/schemas/CategoriaUsuario"
            }
          },
          "type": "object",
          "required": [
            "categoria"
          ],
          "title": "UsuarioAprobacion"
        },
        "UsuarioOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "nombre": {
              "type": "string",
              "title": "Nombre"
            },
            "apellido": {
              "type": "string",
              "title": "Apellido"
            },
            "email": {
              "type": "string",
              "format": "email",
              "title": "Email"
            },
            "domicilio": {
              "type": "string",
              "title": "Domicilio"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            },
            "categoria": {
              "$ref": "#/components/schemas/CategoriaUsuario"
            },
            "estado_registro": {
              "$ref": "#/components/schemas/EstadoRegistro"
            },
            "bloqueado_por_impago": {
              "type": "boolean",
              "title": "Bloqueado Por Impago"
            },
            "fecha_alta": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha Alta"
            }
          },
          "type": "object",
          "required": [
            "id",
            "nombre",
            "apellido",
            "email",
            "domicilio",
            "pais",
            "categoria",
            "estado_registro",
            "bloqueado_por_impago",
            "fecha_alta"
          ],
          "title": "UsuarioOut"
        },
        "UsuarioRegistroEtapa1": {
          "properties": {
            "nombre": {
              "type": "string",
              "maxLength": 80,
              "title": "Nombre"
            },
            "apellido": {
              "type": "string",
              "maxLength": 80,
              "title": "Apellido"
            },
            "email": {
              "type": "string",
              "format": "email",
              "title": "Email"
            },
            "doc_frente_url": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doc Frente Url"
            },
            "doc_dorso_url": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doc Dorso Url"
            },
            "domicilio": {
              "type": "string",
              "title": "Domicilio"
            },
            "pais": {
              "type": "string",
              "title": "Pais"
            }
          },
          "type": "object",
          "required": [
            "nombre",
            "apellido",
            "email",
            "domicilio",
            "pais"
          ],
          "title": "UsuarioRegistroEtapa1"
        },
        "UsuarioRegistroEtapa2": {
          "properties": {
            "password": {
              "type": "string",
              "minLength": 8,
              "title": "Password"
            }
          },
          "type": "object",
          "required": [
            "password"
          ],
          "title": "UsuarioRegistroEtapa2",
          "description": "El postor ingresa la app y genera su clave personal."
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
...
      }
    }
  }